+++ storage/buffer/src/main/java/io/atomix/storage/buffer/BufferOutput.java
@@ -51,7 +51,7 @@ public interface BufferOutput<T extends BufferOutput<?>> extends AutoCloseable {
    * @param length The number of bytes from the provided byte array to write to the buffer.
    * @return The written buffer.
    */
-  T write(Bytes bytes, long offset, long length);
+  T write(Bytes bytes, int offset, int length);
 
   /**
    * Writes an array of bytes to the buffer.
@@ -61,7 +61,7 @@ public interface BufferOutput<T extends BufferOutput<?>> extends AutoCloseable {
    * @param length The number of bytes from the provided byte array to write to the buffer.
    * @return The written buffer.
    */
-  T write(byte[] bytes, long offset, long length);
+  T write(byte[] bytes, int offset, int length);
 
   /**
    * Writes a buffer to the buffer.
